Uncommenting and fixing remaining bugs
Implement meets constraint in TaskNetwork

Compilable version.
Pending work: implement meets, decide whether ActivityReference should use SerializedValue or ProfileExpressoin
Need to resolve the //TODOs in the code
Need to reorganize NexusMetaSolver to address the order between calls to scheduler and decomposer and to solve conflicts regardless of whether they come from a goal or from a SchedulingActivity

Compilable version.
Pending work: implement meets, decide whether ActivityReference should use SerializedValue or ProfileExpressoin
Need to resolve the //TODOs in the code
Need to reorganize NexusMetaSolver to address the order between calls to scheduler and decomposer and to solve conflicts regardless of whether they come from a goal or from a SchedulingActivity

Mergin

Code generation 2

Code generation

Rebase in progress.
Branch up to date with develop.
Planning classes moved. Parsing and code generation are not

Rebase in progress.
Branch up to date with develop.
Planning classes moved. Parsing and code generation are not